WIP LSP client
This contains the start of an LSP client.
Nothing (except starting the LSP server) works
at the moment and the feature is disabled by default.

+++ src/dged/command.c
@@ -79,7 +79,7 @@ void command_ctx_push_arg(struct command_ctx *ctx, const char *argv) {
 }
 
 void command_ctx_free(struct command_ctx *ctx) {
-  for (uint32_t i = 0; i < ctx->saved_argc; ++i) {
+  for (uint32_t i = 0; i < (uint32_t)ctx->saved_argc; ++i) {
     free((char *)ctx->saved_argv[i]);
   }
